fix check error
