+++ eventmesh-common/build.gradle
@@ -26,7 +26,7 @@ dependencies {
     api "org.apache.commons:commons-lang3"
 
     api "com.jayway.jsonpath:json-path"
-//    implementation 'commons-net:commons-net:3.9.0'
+
     api "io.cloudevents:cloudevents-core"
     api "io.cloudevents:cloudevents-json-jackson"
