+++ eventmesh-transformer/src/main/java/org/apache/eventmesh/transformer/Transformer.java
@@ -15,10 +15,18 @@
  * limitations under the License.
  */
 
-package org.apache.eventmesh.transform;
+package org.apache.eventmesh.transformer;
 
-public enum TransformerType {
-    ORIGINAL,
-    CONSTANT,
-    TEMPLATE
+import com.fasterxml.jackson.core.JsonProcessingException;
+
+/**
+ * EventMesh transformer interface, specified transformer implementation includes:
+ * 1. Constant
+ * 2. Original
+ * 3. Template
+ */
+public interface Transformer {
+
+    String transform(String json) throws JsonProcessingException;
+
 }
