+++ eventmesh-transformer/src/main/java/org/apache/eventmesh/transformer/TransformerBuilder.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.eventmesh.transform;
+package org.apache.eventmesh.transformer;
 
 public class TransformerBuilder {
