+++ settings.gradle
@@ -24,7 +24,7 @@ include 'eventmesh-starter'
 include 'eventmesh-examples'
 include 'eventmesh-spi'
 include 'eventmesh-filter'
-include 'eventmesh-transform'
+include 'eventmesh-transformer'
 
 include 'eventmesh-openconnect:eventmesh-openconnect-java'
 include 'eventmesh-openconnect:eventmesh-openconnect-offsetmgmt-plugin:eventmesh-openconnect-offsetmgmt-api'
